增加小狗魔免相关的代码，一些UI的代码

+++ Assets/Scripts/Client/GameLogic/Abilities/AbilityEventElements.cs
@@ -55,7 +55,7 @@ namespace Demo.GameLogic.Abilities
     class AttackSpeed
     {
         public EventTarget target;
-        public float value;
+        public int value;
     }
 
     // 魔法免疫
@@ -63,7 +63,7 @@ namespace Demo.GameLogic.Abilities
     class MagicImmune
     {
         public EventTarget target;
-        public float value;
+        public int value;
     }
 
     // 位移
